+++ TimothE/Camera.cpp
@@ -87,21 +87,21 @@ void Camera::RecalculateViewMatrix()
 void Camera::PollInput()
 {
 	if (Input::IsKeyDown(KEY_W) || Input::IsKeyHeld(KEY_W)) {
-		float yOffset = cos(glm::radians(_rotation)) * _cameraSpeed * Time::GetDeltaTime();
-		float xOffset = -sin(glm::radians(_rotation)) * _cameraSpeed * Time::GetDeltaTime();
+		float yOffset = (float)cos(glm::radians(_rotation)) * _cameraSpeed * Time::GetDeltaTime();
+		float xOffset = (float)-sin(glm::radians(_rotation)) * _cameraSpeed * Time::GetDeltaTime();
 		_cameraPos.x += xOffset;
 		_cameraPos.y += yOffset;
 	}
 	if (Input::IsKeyDown(KEY_S) || Input::IsKeyHeld(KEY_S)) {
-		_cameraPos.x -= -sin(glm::radians(_rotation)) * _cameraSpeed * Time::GetDeltaTime();
-		_cameraPos.y -= cos(glm::radians(_rotation)) * _cameraSpeed * Time::GetDeltaTime();
+		_cameraPos.x -= (float)-sin(glm::radians(_rotation)) * _cameraSpeed * Time::GetDeltaTime();
+		_cameraPos.y -= (float)cos(glm::radians(_rotation)) * _cameraSpeed * Time::GetDeltaTime();
 	}
 	if (Input::IsKeyDown(KEY_A) || Input::IsKeyHeld(KEY_A)) {
-		_cameraPos.x -= cos(glm::radians(_rotation)) * _cameraSpeed * Time::GetDeltaTime();
-		_cameraPos.y -= sin(glm::radians(_rotation)) * _cameraSpeed * Time::GetDeltaTime();
+		_cameraPos.x -= (float)cos(glm::radians(_rotation)) * _cameraSpeed * Time::GetDeltaTime();
+		_cameraPos.y -= (float)sin(glm::radians(_rotation)) * _cameraSpeed * Time::GetDeltaTime();
 	}
 	if (Input::IsKeyDown(KEY_D) || Input::IsKeyHeld(KEY_D)) {
-		_cameraPos.x += cos(glm::radians(_rotation)) * _cameraSpeed * Time::GetDeltaTime();
-		_cameraPos.y += sin(glm::radians(_rotation)) * _cameraSpeed * Time::GetDeltaTime();
+		_cameraPos.x += (float)cos(glm::radians(_rotation)) * _cameraSpeed * Time::GetDeltaTime();
+		_cameraPos.y += (float)sin(glm::radians(_rotation)) * _cameraSpeed * Time::GetDeltaTime();
 	}
 }
